+++ pystac/extensions/eo.py
@@ -45,14 +45,42 @@ class EOItemExt(ItemExtension):
         Returns:
             List[Band]
         """
-        bands = self.item.properties.get('eo:bands')
-        if bands is None:
-            raise STACError("Missing required field 'eo:bands' in item properties")
-        return [Band(b) for b in bands]
+        return self.get_bands()
 
     @bands.setter
     def bands(self, v):
-        self.item.properties['eo:bands'] = [b.to_dict() for b in v]
+        self.set_bands(v)
+
+    def get_bands(self, asset=None):
+        """Gets an Item or an Asset bands.
+
+        If an Asset is supplied and the bands property exists on the Asset,
+        returns the Asset's value. Otherwise returns the Item's value
+
+        Returns:
+            List[Band]
+        """
+        if asset is not None and 'eo:bands' in asset.properties:
+            bands = asset.properties.get('eo:bands')
+        else:
+            bands = self.item.properties.get('eo:bands')
+
+        if bands is not None:
+            bands = [Band(b) for b in bands]
+
+        return bands
+
+    def set_bands(self, bands, asset=None):
+        """Set an Item or an Asset bands.
+
+        If an Asset is supplied, sets the property on the Asset.
+        Otherwise sets the Item's value.
+        """
+        band_dicts = [b.to_dict() for b in bands]
+        if asset is not None:
+            asset.properties['eo:bands'] = band_dicts
+        else:
+            self.item.properties['eo:bands'] = band_dicts
 
     @property
     def cloud_cover(self):
@@ -62,56 +90,39 @@ class EOItemExt(ItemExtension):
         Returns:
             float or None
         """
-        return self.item.properties.get('eo:cloud_cover')
+        return self.get_cloud_cover()
 
     @cloud_cover.setter
     def cloud_cover(self, v):
-        self.item.properties['eo:cloud_cover'] = v
+        self.set_cloud_cover(v)
+
+    def get_cloud_cover(self, asset=None):
+        """Gets an Item or an Asset cloud_cover.
+
+        If an Asset is supplied and the Item property exists on the Asset,
+        returns the Asset's value. Otherwise returns the Item's value
+
+        Returns:
+            float
+        """
+        if asset is None or 'eo:cloud_cover' not in asset.properties:
+            return self.item.properties.get('eo:cloud_cover')
+        else:
+            return asset.properties.get('eo:cloud_cover')
+
+    def set_cloud_cover(self, cloud_cover, asset=None):
+        """Set an Item or an Asset cloud_cover.
+
+        If an Asset is supplied, sets the property on the Asset.
+        Otherwise sets the Item's value.
+        """
+        if asset is None:
+            self.item.properties['eo:cloud_cover'] = cloud_cover
+        else:
+            asset.properties['eo:cloud_cover'] = cloud_cover
 
     def __repr__(self):
         return '<EOItemExt Item id={}>'.format(self.item.id)
-
-    def get_asset_bands(self, asset):
-        """Gets the bands for the given asset.
-
-        Args:
-            asset [Asset]: The asset from the associated item to get the bands of.
-
-        Returns:
-            List[Band] or None: A list of Band objects associated with the asset, or None
-            if the asset doesn't exist or does not contain band information.
-        """
-        if 'eo:bands' not in asset.properties:
-            return None
-
-        result = []
-        band_length = len(self.bands)
-        for band_idx in asset.properties['eo:bands']:
-            if band_idx >= band_length:
-                raise STACError("Item {} contains assets with band indexes out of range! "
-                                "Band index {} is greater than the number of bands "
-                                "for the item ({})".format(self.item.id, band_idx, band_length))
-            result.append(self.bands[band_idx])
-
-        return result
-
-    def set_asset_bands(self, asset, band_names):
-        """Sets the band information for an asset of this item.
-
-        Args:
-            asset [Asset]: The asset from the associated item to set the bands for.
-            band_names (List[str]): List of band names to assign to the asset.
-                These names must reference names in the item's bands property.
-        """
-        band_idxs = [i for i, b in enumerate(self.bands) if b.name in band_names]
-
-        if len(set(band_idxs)) != len(set(band_names)):
-            missing_band_names = set(band_names) - set([self.bands[i] for i in band_idxs])
-
-            raise KeyError("Bands not found in item's bands: {}".format(','.join(
-                map(lambda x: "'{}'".format(x), missing_band_names))))
-
-        asset.properties['eo:bands'] = band_idxs
 
     @classmethod
     def _object_links(cls):